Extract settings and card builder into libs
