+++ AutoDefineAddon/libs/config.py
@@ -1,0 +1,37 @@
+# Get your unique API key by signing up at http://www.dictionaryapi.com/
+MERRIAM_WEBSTER_API_KEY = "YOUR_KEY_HERE"
+
+MERRIAM_WEBSTER_SPANISH_API_KEY = "YOUR_KEY_HERE"
+# Index of field to insert definitions into (use -1 to turn off)
+DEFINITION_FIELD = 1
+
+# Ignore archaic/obsolete definitions?
+IGNORE_ARCHAIC = True
+
+# Get your unique API key by signing up at http://www.dictionaryapi.com/
+MERRIAM_WEBSTER_MEDICAL_API_KEY = "YOUR_KEY_HERE"
+
+# Open a browser tab with an image search for the same word?
+OPEN_IMAGES_IN_BROWSER = False
+
+# Which dictionary should AutoDefine prefer to get definitions from? Available options are COLLEGIATE and MEDICAL.
+PREFERRED_DICTIONARY = "COLLEGIATE"
+
+# Index of field to insert pronunciations into (use -1 to turn off)
+PRONUNCIATION_FIELD = 0
+
+# Index of field to insert phonetic transcription into (use -1 to turn off)
+PHONETIC_TRANSCRIPTION_FIELD = -1
+
+# Index of field to insert pronunciations into (use -1 to turn off)
+DEDICATED_INDIVIDUAL_BUTTONS = False
+
+PRIMARY_SHORTCUT = "ctrl+alt+e"
+
+DEFINE_ONLY_SHORTCUT = ""
+
+PRONOUNCE_ONLY_SHORTCUT = ""
+
+PHONETIC_TRANSCRIPTION_ONLY_SHORTCUT = ""
+
+PART_OF_SPEECH_ABBREVIATION = {"verb": "v.", "noun": "n.", "adverb": "adv.", "adjective": "adj."}
